Rename /controllers to /controller
