+++ recruitment-api/index.ts
@@ -1,5 +1,5 @@
 import { Application, Router } from "https://deno.land/x/oak/mod.ts";
-import { registerControllers } from "./controllers/index.ts";
+import { registerControllers } from "./controller/index.ts";
 
 const app = new Application();
 const router = new Router();
